feat：为I2C库新增了超时重新初始化的功能
react：重构了部分函数的位置
fix：修复了TCA9555库的一些函数返回值错误
feat：为TCA9555库新增了返回值校验的功能

+++ CH592_Node/APP/include/TCA9555.h
@@ -47,7 +47,7 @@ void TCA_WritePin(const uint8_t device_addr, const uint32_t pin, const uint32_t 
 
 void TCA_TogglePin(const uint8_t device_addr, const uint32_t pin);
 
-uint32_t TCA_ReadPin(const uint8_t device_addr, const uint32_t pin);
+int8_t TCA_ReadPin(const uint8_t device_addr, const uint32_t pin);
 
 void TCA_SetAllPinsInput(const uint8_t device_addr);
 #endif /* TCA9555_H_ */
